requirements.txt
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
dj-database-url==2.2.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pillow==12.1.0
psycopg2-binary==2.9.10
python-telegram-bot==22.6
sqlparse==0.5.5
whitenoise==6.8.2
